QCRAFT-Scheduler/requirements.txt
requests<2.32.3,>=2.31.0
qiskit>=1.1.0,<2.0.0
amazon-braket-sdk>=1.80.0,<2.0.0
qiskit-ibm-runtime>=0.23.0,<1.0.0
qiskit-aer>=0.14.2,<1.0.0
flask>=3.0.3,<4.0.0
pymongo>=4.7.3,<5.0.0
antlr4-python3-runtime==4.9.2
numpy>=1.22.4,<2
python-dotenv>=1.0.1,<2.0.0